+++ NEWS.md
@@ -21,7 +21,9 @@ __New Tune Parameter Ranking and Selection Tools__:
 
 -->
 
-# timetk 2.0.0.9000 (Development Version)
+# timetk 2.1.3.9000 (Development Version)
+
+# timetk 2.1.0 
 
 __New Functions__